# 24001200-km7-ald-BasicBankingSystem-ch5/src/docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Basic Banking Simulation API",
    "description": "This is an API for the challenged issued by Binar Academy, using Swagger as a form of API Documentation for ease of use",
    "contact": {
      "name": "Aldy Ramadhan Syahputra"
    },
    "license": {
      "name": "Free"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090/api/v1"
    }
  ],
  "paths": {
...
    "/accounts": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get All Accounts",
        "description": "Getting All Accounts with no exceptions",
        "operationId": "getAccounts",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully Retrieved All Data"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 6
                          },
                          "bank_name": {
                            "type": "string",
                            "example": "vincenzoo"
                          },
                          "bank_account_number": {
                            "type": "string",
                            "example": "771166"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Creates an Account",
        "description": "Creates a new account with the already created users",
        "operationId": "createAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "balance",
                  "bank_account_number",
                  "bank_name",
                  "user_id"
                ],
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "integer",
                    "example": 8
                  },
                  "bank_name": {
                    "type": "string",
                    "example": "vincenzoo"
                  },
                  "bank_account_number": {
                    "type": "string",
                    "example": "771166"
                  },
                  "balance": {
                    "type": "integer",
                    "example": 200000
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account Created Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Account Created Successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 6
                        },
                        "user_id": {
                          "type": "string",
                          "example": "8"
                        },
                        "bank_name": {
                          "type": "string",
                          "example": "vincenzoo"
                        },
                        "bank_account_number": {
                          "type": "string",
                          "example": "771166"
                        },
                        "balance": {
                          "type": "number",
                          "format": "float",
                          "example": 200000
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed response cases",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "message": {
                      "type": "string",
                      "enum": [
                        "Failed To Create User, Make Sure To fill all Form",
                        "Email must be a Gmail address",
                        "Password must be at least 8 characters long",
                        "Identity number must be a number",
                        "Email Already Used",
                        "Identity Number Already Used"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
    "/accounts/{id}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get Accounts by ID",
        "description": "Get a single account by their ID",
        "operationId": "getAccountById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully Retrieved The Data"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "bank_name": {
                          "type": "string",
                          "example": "vincenzoo"
                        },
                        "bank_account_number": {
                          "type": "string",
                          "example": "771166"
                        },
                        "balance": {
                          "type": "number",
                          "format": "float",
                          "example": 200000
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "There's no Such Account!"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}